% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_roll_up_samples.R
\name{code_sql_roll_up_samples}
\alias{code_sql_roll_up_samples}
\title{Rolls up Sample Years}
\usage{
code_sql_roll_up_samples(years, sample_function)
}
\arguments{
\item{years}{the years to roll up}

\item{sample_function}{function that makes a sample for a single year}
}
\value{
A data frame with distinct cohort records
}
\description{
Selects distinct examinees with the highest score
}
