% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_sql_select_only_highest_score.R
\name{code_sql_select_only_highest_score}
\alias{code_sql_select_only_highest_score}
\title{Selects Highest Score}
\usage{
code_sql_select_only_highest_score(sample_sql, by_exam = FALSE)
}
\arguments{
\item{sql}{SQL code from a cohort sample}
}
\value{
A data frame with distinct cohort records
}
\description{
Selects distinct examinees with the highest score
}
\keyword{database,}
\keyword{location}
\keyword{state,}
