% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{DB Connection}
\usage{
db_connect(database = "test_scores")
}
\arguments{
\item{database}{Prefix to match on}
}
\description{
Connects to a database
}
\keyword{database}

