% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Frequency Distribution}
\usage{
freq(x)
}
\arguments{
\item{x}{the vector of observations}
}
\description{
Returns a frequency distribution with more detail than a simple count of objects.
}
\keyword{frequncy}
\keyword{statistics,}
