% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.sample.standardized.R
\name{is.sample.standardized}
\alias{is.sample.standardized}
\title{Is this Sample Standardized?}
\usage{
is.sample.standardized(df)
}
\arguments{
\item{df}{the data.frame that you want to test}
}
\value{
Boolean indicating whether the data.frame is standarized
}
\description{
If the data.frame passes the test then we know for sure that it will have the expected standard columns that can be used in typical analysis and visualizations. This function looks for these columns in this order: person, year, exam, score, tg, us, location, school, district, sector, race, gender, waiver
}
\details{
Looks to make sure that the data.frame columns meet the requirements for standard visualizations.
}
