% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_states.R
\name{lookup_states}
\alias{lookup_states}
\title{Lookup State Information}
\usage{
lookup_states(...)
}
\description{
Quick lookup function for the states table
}
\keyword{abbreviations}
\keyword{and}
\keyword{descriptions}
\keyword{state}

