% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull Data}
\usage{
pull_data(sql)
}
\arguments{
\item{sql}{The SQL code to be executed}
}
\description{
Pulls data from the active database instance, if none exists this function attempts to connect to Cornerstone using the function database_cornerstone (which you need to code with your credientials)
}
\examples{

}
\keyword{database}

