% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp.R
\name{sdp}
\alias{sdp}
\title{Standard Deviation for Population}
\source{
http://stackoverflow.com/a/30812437/131753
}
\usage{
sdp(x)
}
\arguments{
\item{x}{the vector of observations}
}
\description{
The fact that R is missing this is baffling, but we need this all the time when we are doing descriptive statistics so here it is.
}
\details{
Returns the standard deviation for a population (using N and not N-1 as denominator)
}
\keyword{deviation,}
\keyword{standard}
\keyword{statistics,}
\keyword{variability}
