% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_wiche.R
\name{lookup_wiche}
\alias{lookup_wiche}
\title{Lookup WICHE Enrollment Projections}
\usage{
lookup_wiche(location = "us", sector = "p", race = "all",
  gender = "all", grade = "g", year = NULL)
}
\description{
Quick lookup function for the WICHE table
}
\keyword{participation}
\keyword{rates}

