% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull Data}
\usage{
pull_data(sql, db_alt = NULL)
}
\arguments{
\item{sql}{The SQL code to be executed}

\item{db}{Specify different database (dbr object)}
}
\description{
Pulls data from the active database instance or the provided dbr object.
}
\keyword{database}

